Fix some details
