--- src/App.scss
@@ -1,1 +1,0 @@
-@import 'styles/main.scss';
