+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './styles/main.scss';
-import App from './App';
+import App from './components/App';
 
 ReactDOM.render(<App />, document.getElementById('root'));
