+++ src/styles/App.scss
@@ -1,0 +1,1 @@
+@import './main.scss';
