+++ src/styles/main.scss
@@ -7,3 +7,4 @@
 @import 'components/landing';
 @import 'components/shared_form';
 @import 'components/cardpreview';
+@import 'components/design';
